requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.3
djangorestframework==3.14.0
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
py==1.11.0
pytest==6.2.5
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
